добавлена минимальная логика работы экрана трекера
- Добавлено общение с TrackService для получения актуального времени записи трека
- Добавлены проверки и методы для восстановления экрана трекера при переходе между экранами
- Не работает восстановление экрана после смахивания приложения (необходима доработка)
- Добавлена проверка на NaN для значений

+++ cache/src/main/java/io/github/kaczmarek/cache/room/model/TrackDBModel.kt
@@ -5,12 +5,12 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "track")
 class TrackDBModel(
-        @PrimaryKey
-        val id: String, // Индетификатор трека
-        val distance: Double?, // Пройденная дистанция
-        val duration: Long, // Продолжительность записи
-        val maxSpeed: Double?, // Максимальная скорость
-        val averageSpeed: Double?, // Средняя скорость
-        val currentSpeed: Double?, // Текущая скорость (последняя зафиксированная)
-        val isFinishedRecord: Boolean // Флаг остановлена ли запись по текущему треку
+    @PrimaryKey
+    val id: String, // Индетификатор трека
+    val distance: Float?, // Пройденная дистанция
+    val duration: Long, // Продолжительность записи
+    val maxSpeed: Float?, // Максимальная скорость
+    val averageSpeed: Double?, // Средняя скорость
+    val currentSpeed: Float?, // Текущая скорость (последняя зафиксированная)
+    val isFinishedRecord: Boolean // Флаг остановлена ли запись по текущему треку
 )
